requirements.txt
bddl==3.5.0
cloudpickle==1.6.0
colorama==0.4.6
diffusers==0.29.2
easydict==1.13
einops==0.8.0
fairscale==0.4.13
fire==0.6.0
ftfy==6.2.3
gdown==5.2.0
gym==0.17.3
h5py==3.11.0
huggingface-hub==0.24.5

imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
matplotlib==3.7.5
matplotlib-inline==0.1.7
moviepy==1.0.3
mujoco==3.2.3
numpy==1.24.3
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pipreqs==0.5.0
piq==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
robomimic==0.2.0
robosuite==1.4.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
termcolor==2.4.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
torch==2.4.0
torch-fidelity==0.3.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
typing_extensions==4.12.2
wandb==0.17.5
